US Food Sales dataset
This involved the analysis of a US Food Sales dataset obtained from Kaggle. The primary objective was to extract valuable insights from the data to support decision-making and gain a better understanding of sales patterns in different regions and cities. The project consisted of several key steps, including data preprocessing, visualization, and the creation of interactive dashboards using Power BI.